# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "Applink API Documentation",
    "x-logo": {
      "url": "../images/logo.png"
    }
  },
...
  "paths": {
...
    "/subscription/send": {
      "post": {
        "tags": [
          "subscription/userSubscription"
        ],
        "summary": "User Subscription",
        "description": "",
        "operationId": "userSubscription",
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/userSubscription"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/responseOfUserSubscription"
                }
              }
            }
          }
        },
        "x-code-samples": [
          {
            "lang": "Java",
            "source": "coming soon"
          },
          {
            "lang": "PHP",
            "source": "coming soon"
          },
          {
            "lang": "javaScript",
            "source": "coming soon"
          },
          {
            "lang": ".net",
            "source": "coming soon"
          }
        ]
      }
    },
    "/subscription/query-base": {
      "post": {
        "tags": [
          "subscription/baseSize"
        ],
        "summary": "Base Size",
        "description": "",
        "operationId": "baseSize",
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/baseSize"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/responseOfBaseSize"
                }
              }
            }
          }
        },
        "x-code-samples": [
          {
            "lang": "Java",
            "source": "coming soon"
          },
          {
            "lang": "PHP",
            "source": "coming soon"
          },
          {
            "lang": "javaScript",
            "source": "coming soon"
          },
          {
            "lang": ".net",
            "source": "coming soon"
          }
        ]
      }
    },
    "/subscription/getSubscriberChargingInfo": {
      "post": {
        "tags": [
          "subscription/getSubscriberChargingInfo"
        ],
        "summary": "Get Subscriber Charging Info",
        "description": "",
        "operationId": "getSubscriberChargingInfo",
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/getSubscriberChargingInfo"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/responseOfgetSubscriberChargingInfo"
                }
              }
            }
          }
        },
        "x-code-samples": [
          {
            "lang": "Java",
            "source": "coming soon"
          },
          {
            "lang": "PHP",
            "source": "coming soon"
          },
          {
            "lang": "javaScript",
            "source": "coming soon"
          },
          {
            "lang": ".net",
            "source": "coming soon"
          }
        ]
      }
    },
    "/subscription/notify": {
      "post": {
        "tags": [
          "subscription/notify"
        ],
        "summary": "Subscriber Notifications",
        "description": "",
        "operationId": "subscriberNotify",
        "requestBody": {
          "content": {
            "application/json;charset=utf-8": {
              "schema": {
                "$ref": "#/components/schemas/subscriberNotify"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json;charset=utf-8": {
                "schema": {
                  "$ref": "#/components/schemas/responseOfSubscriberNotify"
                }
              }
            }
          }
        },
        "x-code-samples": [
          {
            "lang": "Java",
            "source": "coming soon"
          },
          {
            "lang": "PHP",
            "source": "coming soon"
          },
          {
            "lang": "javaScript",
            "source": "coming soon"
          },
          {
            "lang": ".net",
            "source": "coming soon"
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.applink.com.bd"
    }
  ],
  "components": {
    "schemas": {
...
      "userSubscription": {
        "type": "object",
        "required": [
          "applicationId",
          "password",
          "subscriberId",
          "action"
        ],
        "properties": {
          "applicationId": {
            "type": "string",
            "example": "APP_999999",
            "description": "Used to identify the application. This is a unique identifier generated while provisioning an application. \n\nOnly a single value can be sent per request."
          },
          "password": {
            "type": "string",
            "example": "95904999aa8edb0c038b3295fdd271de",
            "description": "This API key will be sent to your registered email address on the platform upon application approval."
          },
          "subscriberId": {
            "type": "string",
            "example": "tel:8801975355595",
            "description": "This can be the MSISDN of the subscriber to be charged. This is a unique identifier.\n\ntel: is for MSISDN\n\nSubcriber: tel:8801975355595\n\n Note: tel might be a masked number depending on the type of the application\n\nOnly a single value can be sent per request."
          },
          "action": {
            "type": "string",
            "enum": [
              0,
              1
            ],
            "example": "0",
            "description": " 0 -user unsubscription\n\n 1 -user subscription\n\n "
          }
        }
      },
      "responseOfUserSubscription": {
        "type": "object",
        "required": [
          "version",
          "statusCode",
          "statusDetail",
          "subscriptionStatus"
        ],
        "properties": {
          "version": {
            "type": "string",
            "example": "1.0.",
            "description": "API version shall be numbered as 1.0, 2.0 etc.\n\nIf version is specified in the request, same version must be sent in the response.\n\nIf version is not specified in the request, the latest version will be specified in the response."
          },
          "statusCode": {
            "type": "string",
            "example": "S1000",
            "description": "The status code for the entire request."
          },
          "statusDetail": {
            "type": "string",
            "example": "not registered",
            "description": "Description of the status for the entire request."
          },
          "subscriptionStatus": {
            "type": "string",
            "example": "UNREGISTERED.",
            "description": "Status of the subscription eg:UNREGISTERED,REGISTERED"
          }
        }
      },
      "baseSize": {
        "type": "object",
        "required": [
          "applicationId",
          "password"
        ],
        "properties": {
          "applicationId": {
            "type": "string",
            "example": "APP_000201",
            "description": "Used to identify the application. This is a unique identifier generated while provisioning an application. \n\nOnly a single value can be sent per request."
          },
          "password": {
            "type": "string",
            "example": "39a8d1cb245029d0560619a2b388669c",
            "description": "This API key will be sent to your registered email address on the platform upon application approval."
          }
        }
      },
      "responseOfBaseSize": {
        "type": "object",
        "required": [
          "version",
          "baseSize",
          "statusCode",
          "statusDetail"
        ],
        "properties": {
          "baseSize": {
            "type": "string",
            "example": "0",
            "description": "Number of registered users"
          },
          "version": {
            "type": "string",
            "example": "1.0.",
            "description": "API version shall be numbered as 1.0, 2.0 etc.\n\nIf version is specified in the request, same version must be sent in the response.\n\nIf version is not specified in the request, the latest version will be specified in the response."
          },
          "statusCode": {
            "type": "string",
            "example": "S1000",
            "description": "The status code for the entire request."
          },
          "statusDetail": {
            "type": "string",
            "example": "Success.",
            "description": "Description of the status for the entire request."
          }
        }
      },
      "getSubscriberChargingInfo": {
        "type": "object",
        "required": [
          "applicationId",
          "password",
          "subscriberId"
        ],
        "properties": {
          "applicationId": {
            "type": "string",
            "example": "APP_102672",
            "description": "Used to identify the application. This is a unique identifier generated while provisioning an application. \n\nOnly a single value can be sent per request."
          },
          "password": {
            "type": "string",
            "example": "cf2b9e361c13bc54b86d3c8180b0582fd",
            "description": "This API key will be sent to your registered email address on the platform upon application approval."
          },
          "subscriberIds": {
            "type": "string",
            "example": [
              "tel: 8801973579363",
              "tel: 8801973579363"
            ],
            "description": "MSISDNs of the list of subscribers. Eg: [tel: 8801973579363, tel:8801973579363]\n\n If the application accepts masked numbers, list of\n\n masked numbers must be communicated in this parameter\n\n List can contain maximum of 10 MSISDNs per request"
          }
        }
      },
      "subscriberInfo": {
        "type": "object",
        "required": [
          "subscriberId"
        ],
        "properties": {
          "subscriberId": {
            "type": "string",
            "example": "tel:8801973579363",
            "description": "MSISDN of the subscriber.\n\n If the application accepts masked numbers, \n\n list of masked numbers must be communicated in this parameter\n\n THIS IS A MANDATORY FILED"
          },
          "subscriptionStatus": {
            "type": "string",
            "example": "REGISTERED",
            "description": "Status of the subscription \n\n Eg: INITIAL,\n\n REG_PENDING, TRIAL, REGISTERED,\n\n UNREGISTERED and TEMPORARY_BLOCKED"
          },
          "lastChargedDate": {
            "type": "string",
            "example": "2020-01-23 22:03:22",
            "description": "The successful last charged date of the subscriber\n\n in the format of YYYY-MM-DD hh:mm:ss"
          },
          "lastChargedAmount": {
            "type": "string",
            "example": "30.00 LKR",
            "description": "The successful last charged amount for the\n\n subscriber with the currency code"
          },
          "numberType": {
            "type": "string",
            "example": "postpaid",
            "description": "Denotes whether the subscriber is prepaid or postpaid"
          },
          "statusCode": {
            "type": "string",
            "description": "Success or error code for each subscriber and for\n\n the entire request",
            "example": "S1000"
          },
          "statusDetail": {
            "type": "string",
            "example": "Request was successfully processes"
          }
        }
      },
      "responseOfgetSubscriberChargingInfo": {
        "type": "object",
        "required": [
          "version",
          "statusCode",
          "statusDetail"
        ],
        "properties": {
          "version": {
            "type": "string",
            "example": "1.0",
            "description": "API version shall be numbered as 1.0, 2.0 etc.\n\n If version is specified in the request, same version must be sent in the response.\n\n If version is not specified in the request, the latest version will be specified in the response."
          },
          "destinationResponses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/subscriberInfo"
            }
          },
          "statusCode": {
            "type": "string",
            "example": "S1000",
            "description": "The status code for the entire request"
          },
          "statusDetail": {
            "type": "string",
            "example": "Success.",
            "description": " Below table provides information on different subscription statuses returned \n\n by the getSubscriptionChargingInfo service.\n\n INITIAL - Subscription request is sent to the system. The subscription request\n\n might not be completed may be due to a system error. Not even the charging\n\n call is made during this instance.\n\n Parameters returns in response : subscriberId,subscriptionStatus,\n\n numberType,statusCode,statusDetail\n\n \n\n REG_PENDING - Subscription request is made and charging request is also sent\n\n but charging is not yet success.(Eg: Charging response not yet received from the IN,\n\n Insufficient balance)\n\n Parameters returns in response : subscriberId,subscriptionStatus,\n\n numberType,statusCode,statusDetail\n\n TRIAL - Subscription request is made and the platform allows subscribers to use\n\n the service without charging as they have a free trial period enabled.\n\n Parameters returns in response : subscriberId,subscriptionStatus,\n\n numberType,statusCode,statusDetail\n\n REGISTERED - Subscription request is made, charging is successful and the\n\n subscriber can receive the service.If the application registered to is a free\n\n application lastChargedDate to be omitted and lastChargedAmount to be sent as\n\n 0.00 LKR\n\n Parameters returns in response : subscriberId,subscriptionStatus,\n\n lastChargedDate,lastChargedAmount,numberType,statusCode,statusDetail\n\n UNREGISTERED - Subscription request is made and unsubscription is also\n\n done by the subscriber for the particular service.If the application unregistered\n\n from is a free application lastChargedDate to be omitted and lastChargedAmount to be\n\n sent as 0.00 LKR\n\n Parameters returns in response : subscriberId,subscriptionStatus, lastChargedDate,lastChargedAmount,numberType,statusCode,statusDetail\n\n TEMPORARY_BLOCKED - Subscription request is made initial charging was\n\n also successful and was in the registered status once but\n\n recursive charging has failed\n\n (Eg: Due to insufficient balance) and service is temporarily blocked for this subscriber\n\n If the application registered to is a free\n\n application lastChargedDate to be omitted and lastChargedAmount to be sent as 0.00 LKR\n\n Parameters returns in response : subscriberId,subscriptionStatus,lastChargedDate,lastChargedAmount,\n\n numberType,statusCode,statusDetail\n\n The status detail for the entire request\n\n S1000  - Process completed successfully for all the available destination numbers.\n\n E1303  - IP address from which this request originated is not provisioned to send request to application <application-id>. Please use a provisioned system to send request or contact system admin to provision the new IP.\n\n E1312  - Request is Invalid. <specify_the_reason> Refer the TAP NBL API Developer Guide for the mandatory fields and correct format of the request.\n\n E1313  - Authentication failed. No such active application with applicationId <application-id>, or no active service provider or the given password in the request is invalid.\n\n E1317  - < MSISDN > in request is invalid or not allowed.\n\n E1325  - Format of the address is invalid. Expected format is tel: 94712342345.\n\n E1601  - System experienced an unexpected error.\n\n E1603  - Temporary System Error occurred while delivering your request "
          }
        }
      },
      "subscriberNotify": {
        "type": "object",
        "required": [
          "timeStamp",
          "version",
          "password",
          "applicationId",
          "password",
          "subscriberId",
          "frequency",
          "status"
        ],
        "properties": {
          "timeStamp": {
            "type": "string",
            "example": "20120113082110",
            "description": "The timestamp sent from the SMS.\n\"yyMMddHHmm\"\nyy – last two digits of the year (00-99)\nMM – month (01-12)\ndd – day (01-31)\nHH – hour (00-23)\nmm- minute (00-59)"
          },
          "version": {
            "type": "string",
            "example": "1.0",
            "description": "API version shall be numbered as 1.0, 2.0 etc.\n\nIf version is specified in the request, same version must be sent in the response.\n\nIf version is not specified in the request, the latest version will be specified in the response."
          },
          "applicationId": {
            "type": "string",
            "example": "APP_999999",
            "description": "Used to identify the application. This is a unique identifier generated while provisioning an application. \n\nOnly a single value can be sent per request."
          },
          "password": {
            "type": "string",
            "example": "95904999aa8edb0c038b3295fdd271de",
            "description": "This API key will be sent to your registered email address on the platform upon application approval."
          },
          "subscriberId": {
            "type": "string",
            "example": "tel:8801973579363",
            "description": "This can be the MSISDN of the subscriber to be charged. This is a unique identifier.\n\ntel: is for MSISDN\n\nSubcriber: tel:8801973579363\n\n Note: tel might be a masked number depending on the type of the application\n\nOnly a single value can be sent per request."
          },
          "frequency": {
            "type": "string",
            "enum": [
              "daily",
              "weekly",
              "monthly",
              "yearly"
            ],
            "example": "monthly",
            "description": "Frequency of notifications being sent"
          },
          "status": {
            "type": "string",
            "example": "REGISTERED.",
            "description": "Status of the subscription eg:UNREGISTERED,REGISTERED"
          }
        }
      },
      "responseOfSubscriberNotify": {
        "type": "object",
        "required": [
          "statusCode",
          "statusDetail"
        ],
        "properties": {
          "statusCode": {
            "type": "string",
            "example": "S1000",
            "description": "The status code for the entire request."
          },
          "statusDetail": {
            "type": "string",
            "example": "Request was successfully processed",
            "description": "Description of the status for the entire request."
          }
        }
      },
...
    }
  }
}